@@ -5,19 +5,12 @@ repository: https://github.com/minhnt3/navi_light
 issue_tracker: https://github.com/minhnt3/navi_light/issues
 
 environment:
-  sdk: ">=3.0.5 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
 
 dependencies:
-  _fe_analyzer_shared: ^61.0.0
-  analyzer: ^5.13.0
+  analyzer: ">=5.12.0 <7.0.0"
   analyzer_plugin: ^0.11.2
-  custom_lint_builder: ^0.5.0
-  dart_mappable: ^3.1.0
-  dart_style: ^2.3.0
-  path: ^1.8.0
-  yaml: ^3.1.2
-
-dev_dependencies:
-  build_runner: ^2.4.0
-  dart_mappable_builder: ^3.1.0
-  flutter_lints: ^2.0.0
+  collection: ^1.16.0
+  custom_lint_builder: ">=0.4.0 <0.6.0"
+  glob: ^2.1.2
+  path: ^1.8.1
