@@ -17,15 +17,13 @@ export 'package:analyzer_plugin/protocol/protocol_common.dart'
 export 'package:analyzer_plugin/src/utilities/change_builder/change_builder_dart.dart';
 export 'package:analyzer_plugin/utilities/change_builder/change_builder_dart.dart';
 export 'package:custom_lint_builder/custom_lint_builder.dart';
-export 'package:dart_mappable/dart_mappable.dart';
 export 'package:dart_style/dart_style.dart';
 export 'package:path/path.dart';
-export 'package:yaml/yaml.dart';
 
-export 'base/load_options_mixin.dart';
-export 'base/options_assist.dart';
+export 'base/excludable.dart';
 export 'base/options_fix.dart';
 export 'base/options_lint_rule.dart';
+export 'base/rule_config.dart';
 export 'lints/avoid_unnecessary_async_function.dart';
 export 'lints/incorrect_todo_comment.dart';
 export 'lints/lines_longer_than_maximum_chars.dart';
@@ -34,16 +32,14 @@ export 'lints/prefer_is_not_empty_string.dart';
 export 'lints/prefer_named_parameters.dart';
 export 'model/code_line.dart';
 export 'options/avoid_unnecessary_async_function_option.dart';
-export 'options/base/options_mixin.dart';
-export 'options/hooks/error_severity_hook.dart';
 export 'options/incorrect_todo_comment_option.dart';
 export 'options/lines_longer_than_maximum_chars_option.dart';
-export 'options/options.dart';
 export 'options/prefer_is_empty_string_option.dart';
 export 'options/prefer_is_not_empty_string_option.dart';
 export 'options/prefer_named_parameters_option.dart';
-export 'options/rules_option.dart';
+export 'utils/data_mapper.dart';
 export 'utils/extensions.dart';
+export 'utils/path_utils.dart';
 export 'visitor/recursive_await_expression_visitor.dart';
 export 'visitor/recursive_constructor_and_function_and_method_declaration_visitor.dart';
 export 'visitor/recursive_function_and_method_declaration_visitor.dart';
