@@ -1,38 +1,45 @@
 import '../index.dart';
 
-class LinesLongerThanMaximumChars extends OptionsLintRule {
-  const LinesLongerThanMaximumChars() : super(code: _code);
+class LinesLongerThanMaximumChars extends OptionsLintRule<LinesLongerThanMaximumCharsOption> {
+  LinesLongerThanMaximumChars(
+    CustomLintConfigs configs,
+  ) : super(
+          RuleConfig(
+            name: lintName,
+            configs: configs,
+            paramsParser: LinesLongerThanMaximumCharsOption.fromMap,
+            problemMessage: (_) =>
+                'The line length exceeds the maximum character limit.\nTry breaking the line across multiple lines.',
+          ),
+        );
 
-  static const _code = LintCode(
-    name: 'lines_longer_than_maximum_chars',
-    problemMessage:
-        'The line length exceeds the maximum character limit.\nTry breaking the line across multiple lines.',
-  );
+  static const String lintName = 'lines_longer_than_maximum_chars';
 
   @override
-  void runWithOptions(
+  Future<void> run(
     CustomLintResolver resolver,
     ErrorReporter reporter,
     CustomLintContext context,
-    Options options,
-  ) {
-    if (options.isFileRuleExcluded(resolver.path)) {
+  ) async {
+    final rootPath = await resolver.rootPath;
+    final parameters = config.parameters;
+    if (parameters.shouldSkipAnalysis(
+      path: resolver.path,
+      rootPath: rootPath,
+    )) {
       return;
     }
 
-    final option = options.rules.linesLongerThanMaximumCharsOption;
-
-    if (option.shouldSkipFile(resolver.path)) {
-      return;
-    }
+    final code = this.code.copyWith(
+          errorSeverity: parameters.severity ?? this.code.errorSeverity,
+        );
 
     resolver.getLineContents((codeLine) {
-      if (codeLine.lineLength > option.maximumChars &&
+      if (codeLine.lineLength > parameters.maximumChars &&
           !codeLine.isComment &&
           !codeLine.isImportStatement &&
           !codeLine.isExportStatement &&
           !codeLine.isString) {
-        final code = this.code.copyWith(errorSeverity: option.severity);
         reporter.reportErrorForOffset(code, codeLine.lineOffset, codeLine.lineLength);
       }
     });
@@ -40,20 +47,21 @@ class LinesLongerThanMaximumChars extends OptionsLintRule {
 
   @override
   List<Fix> getFixes() => [
-        FormatCode(),
+        FormatCode(config),
       ];
 }
 
-class FormatCode extends OptionsFix {
+class FormatCode extends OptionsFix<LinesLongerThanMaximumCharsOption> {
+  FormatCode(super.config);
+
   @override
-  void runWithOptions(
+  Future<void> run(
     CustomLintResolver resolver,
     ChangeReporter reporter,
     CustomLintContext context,
     AnalysisError analysisError,
     List<AnalysisError> others,
-    Options options,
-  ) {
+  ) async {
     final changeBuilder = reporter.createChangeBuilder(
       message: 'Format this document',
       priority: 78,
@@ -61,7 +69,7 @@ class FormatCode extends OptionsFix {
     changeBuilder.addDartFileEdit((builder) {
       builder.formatWithPageWidth(
         resolver.documentRange,
-        pageWidth: options.rules.linesLongerThanMaximumCharsOption.maximumChars,
+        pageWidth: config.parameters.maximumChars,
       );
     });
   }
